Moved get model definition
