+++ aviary/interface/download_models.py
@@ -2,52 +2,7 @@ import os
 from pathlib import Path
 import argparse
 import shutil
-from aviary.utils.functions import get_aviary_resource_path
-
-
-def get_model(file_name: str, verbose=False) -> Path:
-    '''
-    This function attempts to find the path to a file or folder in aviary/models
-    If the path cannot be found in any of the locations, a FileNotFoundError is raised.
-
-    Parameters
-    ----------
-    path : str or Path
-        The input path, either as a string or a Path object.
-
-    Returns
-    -------
-    aviary_path
-        The absolute path to the file.
-
-    Raises
-    ------
-    FileNotFoundError
-        If the path is not found.
-    '''
-
-    # Get the path to Aviary's models
-    path = Path('models', file_name)
-    aviary_path = Path(get_aviary_resource_path(str(path)))
-
-    # If the file name was provided without a path, check in the subfolders
-    if not aviary_path.exists():
-        sub_dirs = [x[0] for x in os.walk(get_aviary_resource_path('models'))]
-        for sub_dir in sub_dirs:
-            temp_path = Path(sub_dir, file_name)
-            if temp_path.exists():
-                # only return the first matching file
-                aviary_path = temp_path
-                continue
-
-    # If the path still doesn't exist, raise an error.
-    if not aviary_path.exists():
-        raise FileNotFoundError(
-            f"File or Folder not found in Aviary's hangar"
-        )
-    if verbose:
-        print('found', aviary_path, '\n')
-    return aviary_path
+from aviary.utils.functions import get_model
 
 
 def save_file(aviary_path: Path, outdir: Path, verbose=False) -> Path:
